App Environment to provide unique worker and cell instance names
Instead of the MachineName (which was explicitly mapped to the machine's
host name, IHostContext now allows to provide an arbitrary WorkerName
(which could still be the machine host name).

IHostContext now should also provide a unique worker instance name which
is supposed to be unique for each run and worker (hence "unique instance").

Additionally, IHostContext defines a method to generate similar unique
cell instance names on demand (whenever a new cell is created).

IApplicationEnvironment provides access to all these fields.